Add graph sequence representation vector option
In earlier version, we use the last graph from the last sentence
to compute the aggregate representation. However, it may be useful
to obtain information from all sentence graphs. This adds an option
to enable such behavior, by computing an aggregate for each graph
and then creating a summary of the vector sequence using a GRU.